+++ src/components/layout/pageBannerPPT/PageBannerPPT.tsx
@@ -56,6 +56,9 @@ export default function PageBannerPPT({ pageNo, backPage }: PageBannerPPTProps) 
 
   const loggedIn = isLoggedIn();
 
+  const getAccess = () => application.content4 as ProposalAccess[];
+  const getProposal = () => application.content2 as Proposal;
+
   const isDisableEndpoints = () => {
     /* c8 ignore start */
     const testDefaultUser = window.localStorage.getItem('cypress:defaultUserLoggedIn') === 'true';
@@ -64,9 +67,6 @@ export default function PageBannerPPT({ pageNo, backPage }: PageBannerPPTProps) 
     } /* c8 ignore end */
     return !loggedIn;
   };
-
-  const getAccess = () => application.content4 as ProposalAccess[];
-  const getProposal = () => application.content2 as Proposal;
 
   const validateTooltip = () => {
     return 'validationBtn.tooltip';